refactor: 서비스 컨버터 로직 분리 #19
